requirements.txt
# Core dependencies
python-binance==1.0.19
ccxt==4.2.25
pandas==2.2.0
numpy==1.26.3
ta==0.11.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Configuration
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1

# Reporting
jinja2==3.1.3
weasyprint==60.2

# Notifications
requests==2.31.0
python-telegram-bot==20.8

# Utilities
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
types-requests==2.31.0.20240125

# Async support
aiohttp==3.9.3
asyncio==3.4.3

# Performance
numba==0.59.0